Fix ruff findings across app / tests / evals
Config:
- Ignore B008 (FastAPI `Depends(...)` in defaults is idiomatic).
- Ignore RUF001/RUF002 (Japanese copy deliberately uses full-width `×`).
- Declare api/render/evals as known-first-party so isort groups them
  below third-party imports.

Code:
- Drop unused imports (`Column`, `json`, `Blueprint`).
- Unquote forward-refs (`BlueprintRow`, `LLMClient`) now safe under
  `from __future__ import annotations`.
- Remove redundant `int(round(...))` — `round()` already returns int.
- Organize imports in figure_renderers/__init__.py.
- Drop stale `# noqa` directives that ruff no longer needed.
- Rename unused unpacked var to `_` in test_template_classifier.

+++ app/render/shapes.py
@@ -16,12 +16,12 @@ EMU_PER_INCH: Final[int] = 914400
 
 def inch(v: float) -> int:
     """Convert inches to integer EMU."""
-    return int(round(v * EMU_PER_INCH))
+    return round(v * EMU_PER_INCH)
 
 
 def _i(v: float) -> int:
     """Integer guard for EMU values (PowerPoint compatibility)."""
-    return int(round(v))
+    return round(v)
 
 
 @dataclass(frozen=True)
